Fix a trivial JS console error on most page loads.
The trydotnet.js script is sourced on every page load (see
Views/Shared/_Layout.cshtml), but only needs to do any work on those
that include a Try .NET snippet (e.g. /userguide/recipes).

Most of what it runs unconditionally has no effect (e.g. adding a
'message' event listener), but it was attempting to set a click event
listener on the 'trydotnet-run' element, and trivially failing with a
"Cannot read properties of null" error if that element didn't exist.